refactor(presto): Extract and compile generator
