+++ sqlglot/dialects/dune.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 
-from sqlglot import exp
 from sqlglot.dialects.trino import Trino
+from sqlglot.generators.dune import DuneGenerator
 from sqlglot.parsers.dune import DuneParser
 
 
@@ -12,8 +12,4 @@ class Dune(Trino):
     class Tokenizer(Trino.Tokenizer):
         HEX_STRINGS = ["0x", ("X'", "'")]
 
-    class Generator(Trino.Generator):
-        TRANSFORMS = {
-            **Trino.Generator.TRANSFORMS,
-            exp.HexString: lambda self, e: f"0x{e.this}",
-        }
+    Generator = DuneGenerator
